Fix price stuck loading forever on homepage hairstyle cards
formatPrice() used int.parse() on the amount string, but MySQL
DECIMAL(10,2) columns serialize as strings like "25000.00" - int.parse
throws on the decimal point. The FutureBuilder only checked
!snapshot.hasData to show the spinner, never handling the error state,
so every card's price spun forever instead of erroring visibly.

+++ lib/src/functions/moneyFormatter.dart
@@ -1,7 +1,7 @@
 import 'package:intl/intl.dart';
 
 Future<String> formatPrice(String number, String currencySymbol) async {
-  int price = int.parse(number);
+  double price = double.parse(number);
   String formattedPrice = NumberFormat('#,###').format(price);
   return 'Amount: '+formattedPrice + currencySymbol;
 }
